feat: 리뷰추천수, 완료된 상담횟수 업데이트 구현
#203
